разбила на два main, исправила все замечания
