+++ .gitignore
@@ -1,3 +1,4 @@
 README.md
 /cmake-build-debug
 .idea
+a.out
