+++ test/test_func.c
@@ -7,52 +7,46 @@
 
 void test()
 {
-	int len = 0;
-	char *str;
+	int size = 10;
+	float tests[]={15, 987, 100, 99, 3999, 174, 1, 5.0, 3333, 10};
+    char *answ[]={"XV", "CMLXXXVII", "C", "XCIX", "MMMCMXCIX","CLXXIV", "I","V", "MMMCCCXXXIII", "X"};
+    int len_answ[]={2,9, 1, 4, 9, 6, 1, 1, 12, 1};
 
-	//roman правильно переводит из десятичной системы счисления в римскую
-	assert(strcmp(str = ft_main(1, &len, 15), "XV") == 0);
-	free(str);
-	assert(strcmp(str = ft_main(1, &len, 987), "CMLXXXVII") == 0);
-    free(str);
-	assert(strcmp(str = ft_main(1, &len, 100), "C") == 0);
-    free(str);
-	assert(strcmp(str = ft_main(1, &len, 99), "XCIX") == 0);
-    free(str);
-	assert(strcmp(str = ft_main(1, &len, 3999), "MMMCMXCIX") == 0);
-    free(str);
-	assert(strcmp(str = ft_main(1, &len, 174), "CLXXIV") == 0);
-    free(str);
-	assert(strcmp(str = ft_main(1, &len, 1), "I") == 0);
-    free(str);
-	assert(strcmp(str = ft_main(1, &len, 5.0), "V") == 0);
-    free(str);
-	assert(strcmp(str = ft_main(1, &len, 3333), "MMMCCCXXXIII") == 0);
-    free(str);
+    float wrong_tests[]={-15, 0, 4000, 5.2, -7.008, 8008, 1234567, -0, 0.0, -1};
 
-	//roman правильно реагирует на некорректные значения
-	assert(strcmp(ft_main(1, &len, -15), "WRONG_NUMBER") == 0);
-	assert(strcmp(ft_main(1, &len, 0), "WRONG_NUMBER") == 0);
-	assert(strcmp(ft_main(1, &len, 4000), "WRONG_NUMBER") == 0);
-	assert(strcmp(ft_main(1, &len, 5.2), "WRONG_NUMBER") == 0);
-	assert(strcmp(ft_main(1, &len, -7.008), "WRONG_NUMBER") == 0);
-	assert(strcmp(ft_main(1, &len, 8008), "WRONG_NUMBER") == 0);
-//	assert(strcmp(ft_main(1, &len, "dgdgdr"), "WRONG_NUMBER") == 0);
-//	assert(strcmp(ft_main(1, &len, "487"), "WRONG_NUMBER") == 0);
+
+    //roman правильно переводит из десятичной системы счисления в римскую
+	for (int i=0; i< size; i++)
+    {
+        int len = 0;
+        char *str = NULL;
+	    assert(strcmp(str = ft_main(1, &len, tests[i]), answ[i]) == 0);
+	    free(str);
+    }
+    //roman правильно реагирует на некорректные значения
+    for (int i=0; i< size; i++)
+    {
+        int len = 0;
+        assert(strcmp(ft_main(1, &len, wrong_tests[i]), "WRONG_NUMBER") == 0);
+    }
 
     //проверка на длину корректной выходной строки
-    ft_main(1, &len, 3333);
-    free(str);
-    assert(len == 12);
-    ft_main(1, &len, 1);
-    free(str);
-    assert(len == 1);
+    for (int i=0; i < size; i++)
+    {
+        char *str;
+        int len = 0;
+        str = ft_main(1, &len, tests[i]);
+        free(str);
+        assert(len == len_answ[i]);
+    }
 
     //проверка на длину некорректной выходной строки
-    ft_main(1, &len, 5000);
-    assert(len == 0);
-    ft_main(1, &len, -18);
-    assert(len == 0);
+    for (int i=0; i<size; i++)
+    {
+        int len = 0;
+        ft_main(1, &len, wrong_tests[i]);
+        assert(len == 0);
+    }
 
 	printf("ok!\n");
 }
